requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
weasyprint==62.0
python-dateutil==2.8.2
